+++ Engine/Rendering/CabotEngine/Graphics/VertexBuffer.h
@@ -12,7 +12,7 @@ class VertexBuffer
 
 public:
     VertexBuffer(const Mesh *p_init_data); // コンストラクタでバッファを生成
-    VertexBuffer(UINT num_vertices, const Vertex *p_init_data);
+    VertexBuffer(size_t num_vertices, const Vertex *p_init_data);
     D3D12_VERTEX_BUFFER_VIEW *View(); // 頂点バッファビューを取得
     bool IsValid(); // バッファの生成に成功したかを取得
